weather-agent-project/pyproject.toml
[project]
name = "weather-agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
